+++ hubdle-settings-gradle-plugin/main/kotlin/com/javiersc/hubdle/HubdleSettingsExtension.kt
@@ -1,0 +1,18 @@
+package com.javiersc.hubdle
+
+import com.javiersc.hubdle.extensions.AutoIncludeExtension
+import javax.inject.Inject
+import org.gradle.api.Action
+import org.gradle.api.model.ObjectFactory
+import org.gradle.kotlin.dsl.newInstance
+
+@HubdleSettingsDslMarker
+public open class HubdleSettingsExtension @Inject constructor(objects: ObjectFactory) {
+
+    public val autoInclude: AutoIncludeExtension = objects.newInstance()
+
+    @HubdleSettingsDslMarker
+    public fun autoInclude(action: Action<AutoIncludeExtension> = Action {}) {
+        action.execute(autoInclude)
+    }
+}